# backend/openapi/v1.json
{
  "components": {
    "schemas": {
...
      "DeclareRootFolder": {
        "additionalProperties": false,
        "description": "Request to declare a destination the Arr does not report.",
        "properties": {
          "instance_id": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        },
        "required": [
          "instance_id",
          "path"
        ],
        "type": "object"
      },
      "Declared": {
        "description": "A destination as declared.",
        "properties": {
          "id": {
            "type": "string"
          },
          "path": {
            "description": "As stored: closed by the separator the Arr writes.",
            "type": "string"
          },
          "verified": {
            "description": "The Arr reported the folder when it was declared. Unverified, it is\ntried at the next sync.",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "path",
          "verified"
        ],
        "type": "object"
      },
      "Deleted": {
        "description": "What a removal answers.",
        "properties": {
          "deleted": {
            "type": "boolean"
          }
        },
        "required": [
          "deleted"
        ],
        "type": "object"
      },
      "ErrorResponse": {
        "description": "The envelope every refusal and failure answers.",
        "properties": {
          "answerable": {
            "description": "Beside `confirm`: whether this caller may send the name back.",
            "type": [
              "boolean",
              "null"
            ]
          },
          "confirm": {
            "description": "The guardrail that asks, for `confirmation_required`. Sending it back\nin `confirm` accepts that question and no other.",
            "type": [
              "string",
              "null"
            ]
          },
          "error": {
            "description": "A stable code: `bad_request`, `unauthorized`, `forbidden`, `not_found`,\n`method_not_allowed`, `conflict`, `confirmation_required`,\n`payload_too_large`, `unsupported_media_type`, `too_many_requests`,\n`busy`, `reauthentication_required`, `passphrase_required`,\n`external_api_error` or `internal_error`. An open list.",
            "type": "string"
          },
          "includes": {
            "description": "Beside `confirm`: the other guardrails the question states, sent back\nin `confirm` with it to accept the question whole.",
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "message": {
            "description": "A sentence for a person, in the interface language. Never part of the\ncontract.",
            "type": "string"
          },
          "reason": {
            "description": "Why a `conflict` refuses, an open list. A job still running, with a\n`Retry-After` header: `apply_running`, `simulation_running`,\n`previews_full`, `sync_running`, `enrichment_running`, `backup_running`,\n`backup_just_taken`, `restore_staging`, `purge_running`. A state that\nstays until somebody changes it: `name_taken`, `in_use`,\n`already_listed`, `not_declared`, `category_mapped`,\n`proposals_replaced`, `dry_run`, `not_cancellable`, `key_pinned`,\n`only_credential`, `secret_unreadable`.",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "error",
          "message"
        ],
        "type": "object"
      },
...
      "RootFolder": {
        "description": "A root folder synchronized from an Arr instance.",
        "properties": {
          "accessible": {
            "type": "boolean"
          },
          "arr_id": {
            "description": "The folder's id in the Arr, `null` for a folder Routarr declares.",
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "category": {
            "type": [
              "string",
              "null"
            ]
          },
          "free_space": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "instance_id": {
            "type": "string"
          },
          "last_accessible_at": {
            "description": "When the folder last answered, as opposed to when it was last seen in a\npass. `null` for one that has never answered.",
            "type": [
              "string",
              "null"
            ]
          },
          "last_synced_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "origin": {
            "description": "`arr` for a folder the instance reports, `declared` for one typed into\nRoutarr. Only the second may be deleted here.",
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "instance_id",
          "path",
          "accessible",
          "origin"
        ],
        "type": "object"
      },
      "RootFolderWithInstance": {
        "allOf": [
          {
            "$ref": "#/components/schemas/RootFolder"
          },
          {
            "properties": {
              "instance_name": {
                "type": "string"
              },
              "instance_type": {
                "type": "string"
              }
            },
            "required": [
              "instance_name",
              "instance_type"
            ],
            "type": "object"
          }
        ],
        "description": "Root folder with instance info for display."
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "in": "header",
        "name": "X-Api-Key",
        "type": "apiKey"
      },
      "bearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
...
    "license": {
      "identifier": "GPL-3.0-only",
      "name": "GPL-3.0-only"
    },
    "title": "Routarr API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/root-folders": {
      "get": {
        "description": "The folders each Arr reports and the destinations declared beside them,\nwith their free space, reachability and category.",
        "operationId": "list_root_folders",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RootFolderWithInstance"
                  },
                  "type": "array"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "A refusal or a failure, in the envelope every error shares."
          }
        },
        "summary": "The root folders, and the category each leads to",
        "tags": [
          "categories"
        ],
        "x-routarr-scope": "read"
      },
      "post": {
        "description": "Written as the Arr sees it: `/data/movies/4k` in a Linux container,\n`D:\\Movies\\4K` or `\\\\nas\\films` on Windows.",
        "operationId": "declare_destination",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeclareRootFolder"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Declared"
                }
              }
            },
            "headers": {
              "Location": {
                "description": "Where the new destination is found.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "A refusal or a failure, in the envelope every error shares."
          }
        },
        "summary": "Declare a destination the Arr does not list",
        "tags": [
          "categories"
        ],
        "x-routarr-scope": "configure"
      }
    },
...
    "/root-folders/{id}": {
      "delete": {
        "description": "Refused with 409 for a folder an Arr reports: the next sync would bring it\nback, without its category.",
        "operationId": "remove_destination",
        "parameters": [
          {
            "description": "The folder's id.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Deleted"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "A refusal or a failure, in the envelope every error shares."
          }
        },
        "summary": "Remove a declared destination",
        "tags": [
          "categories"
        ],
        "x-routarr-scope": "configure"
      }
    },
...
  },
  "security": [
    {
      "api_key": []
    },
    {
      "bearer": []
    }
  ],
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
}